Infix operators

+++ src/main/java/com/jetbrains/jetpad/vclang/term/PrettyPrintable.java
@@ -3,5 +3,5 @@ package com.jetbrains.jetpad.vclang.term;
 import java.util.List;
 
 public interface PrettyPrintable {
-  void prettyPrint(StringBuilder builder, List<String> names, int prec);
+  void prettyPrint(StringBuilder builder, List<String> names, byte prec);
 }
